+++ src/test/java/com/github/kaklakariada/fritzbox/login/Md5ServiceTest.java
@@ -17,32 +17,32 @@
  */
 package com.github.kaklakariada.fritzbox.login;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  * Unit test for {@link Md5Service}
  */
-public class Md5ServiceTest {
+class Md5ServiceTest {
 
     @Test
-    public void testMd5EmptyString() {
+    void testMd5EmptyString() {
         assertMd5Sum("", "d41d8cd98f00b204e9800998ecf8427e");
     }
 
     @Test
-    public void testMd5Test() {
+    void testMd5Test() {
         assertMd5Sum("test", "c8059e2ec7419f590e79d7f1b774bfe6");
     }
 
     @Test
-    public void testMd5Umlauts() {
+    void testMd5Umlauts() {
         assertMd5Sum("1234567z-äbc", "9e224a41eeefa284df7bb0f26c2913e2");
     }
 
     @Test
-    public void testMd5SpecialChars() {
+    void testMd5SpecialChars() {
         assertMd5Sum("!\"§$%&/()=?ßüäöÜÄÖ-.,;:_`´+*#'<>≤|", "ad44a7cb10a95cb0c4d7ae90b0ff118a");
     }