+++ src/test/java/com/github/kaklakariada/fritzbox/model/homeautomation/StatisticsTest.java
@@ -17,61 +17,61 @@
  */
 package com.github.kaklakariada.fritzbox.model.homeautomation;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 import java.util.Optional;
 
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 
-public class StatisticsTest {
-    
-    private static Statistics  statistics;
-    
-    @BeforeClass
-    public static void setupStatistics() {
+class StatisticsTest {
+
+    private static Statistics statistics;
+
+    @BeforeAll
+    static void setupStatistics() {
         statistics = new Statistics();
         statistics.setMeasurementUnit(MeasurementUnit.POWER);
     }
-    
+
     @Test
-    public void computeValueTest() {
-        assertEquals("Test null", null, statistics.computeValue(null));
-        assertEquals("Test Integer", Double.valueOf(0.02), statistics.computeValue("2"));
-        assertEquals("Test Double", null, statistics.computeValue("1.1"));
-        assertEquals("Test Double", null, statistics.computeValue("1.1D"));
+    void computeValueTest() {
+        assertEquals(null, statistics.computeValue(null), "Test null");
+        assertEquals(Double.valueOf(0.02), statistics.computeValue("2"), "Test Integer");
+        assertEquals(null, statistics.computeValue("1.1"), "Test Double");
+        assertEquals(null, statistics.computeValue("1.1D"), "Test Double");
     }
-    
+
     @Test
-    public void getCsvValuesTest() {
+    void getCsvValuesTest() {
         statistics.setCsvValues(null);
         List<Optional<Number>> result = statistics.getValues();
-        assertEquals("(1) Number of entries", 0, result.size());
-        
+        assertEquals(0, result.size(), "(1) Number of entries");
+
         statistics.setCsvValues("");
         result = statistics.getValues();
-        assertEquals("(2) Number of entries", 1, result.size());
-        assertEquals("(2) First number", Optional.empty(), result.get(0));
-        
+        assertEquals(1, result.size(), "(2) Number of entries");
+        assertEquals(Optional.empty(), result.get(0), "(2) First number");
+
         statistics.setCsvValues("1111, 22222");
         result = statistics.getValues();
-        assertEquals("(3) Number of entries", 2, result.size());
-        assertEquals("(3) First number", Double.valueOf(11.11), result.get(0).get());
-        
+        assertEquals(2, result.size(), "(3) Number of entries");
+        assertEquals(Double.valueOf(11.11), result.get(0).get(), "(3) First number");
+
         statistics.setCsvValues("1111,,22222,-");
         result = statistics.getValues();
-        assertEquals("(4) Number of entries", 4, result.size());
-        assertEquals("(4) First number", Double.valueOf(11.11), result.get(0).get());
-        assertEquals("(4) Second number is empty", true, result.get(1).isEmpty());
-        assertEquals("(4) Third number", Double.valueOf(222.22), result.get(2).get());
-        assertEquals("(4) Fourth number is empty", true, result.get(3).isEmpty());
-        
+        assertEquals(4, result.size(), "(4) Number of entries");
+        assertEquals(Double.valueOf(11.11), result.get(0).get(), "(4) First number");
+        assertEquals(true, result.get(1).isEmpty(), "(4) Second number is empty");
+        assertEquals(Double.valueOf(222.22), result.get(2).get(), "(4) Third number");
+        assertEquals(true, result.get(3).isEmpty(), "(4) Fourth number is empty");
+
         statistics.setCsvValues(",1111,-");
         result = statistics.getValues();
-        assertEquals("(5) Number of entries", 3, result.size());
-        assertEquals("(5) First number is empty", true, result.get(0).isEmpty());
-        assertEquals("(5) Second number", Double.valueOf(11.11), result.get(1).get());
-        assertEquals("(5) Third number is empty", true, result.get(2).isEmpty());
+        assertEquals(3, result.size(), "(5) Number of entries");
+        assertEquals(true, result.get(0).isEmpty(), "(5) First number is empty");
+        assertEquals(Double.valueOf(11.11), result.get(1).get(), "(5) Second number");
+        assertEquals(true, result.get(2).isEmpty(), "(5) Third number is empty");
     }
 }