+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [#29](https://github.com/kaklakariada/fritzbox-java-api/issues/29) Add support for login via Pbkdf2 challenge-response (propsed by [@linnex81](https://github.com/linnex81))
 
+### Refactoring
+
+* [#50](https://github.com/kaklakariada/fritzbox-java-api/pull/50) Migrate to Junit 5
+
 ## [1.6.0] - 2021-12-04
 
 ### Added